Cxf-soap tests: Refactor and split
