+++ integration-test-groups/cxf-soap/cxf-soap-client/src/main/resources/application.properties
