+++ integration-test-groups/cxf-soap/cxf-soap-ws-security/src/main/resources/wsdl/UsernameToken.wsdl
@@ -1,0 +1,65 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<definitions targetNamespace="http://www.jboss.org/jbossws/ws-extensions/wssecuritypolicy" name="UsernameToken"
+             xmlns:tns="http://www.jboss.org/jbossws/ws-extensions/wssecuritypolicy"
+             xmlns:xsd="http://www.w3.org/2001/XMLSchema"
+             xmlns:soap="http://schemas.xmlsoap.org/wsdl/soap/"
+             xmlns="http://schemas.xmlsoap.org/wsdl/"
+             xmlns:wsp="http://www.w3.org/ns/ws-policy"
+             xmlns:wsam="http://www.w3.org/2007/05/addressing/metadata"
+             xmlns:wsu="http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-utility-1.0.xsd"
+             xmlns:wsaws="http://www.w3.org/2005/08/addressing"
+             xmlns:sp="http://docs.oasis-open.org/ws-sx/ws-securitypolicy/200702"
+             xmlns:t="http://docs.oasis-open.org/ws-sx/ws-trust/200512">
+    <types>
+        <xsd:schema>
+            <xsd:import namespace="http://www.jboss.org/jbossws/ws-extensions/wssecuritypolicy" schemaLocation="UsernameToken_schema1.xsd"/>
+        </xsd:schema>
+    </types>
+    <message name="sayHello">
+        <part name="parameters" element="tns:sayHello"/>
+    </message>
+    <message name="sayHelloResponse">
+        <part name="parameters" element="tns:sayHelloResponse"/>
+    </message>
+    <portType name="ServiceIface">
+        <operation name="sayHello">
+            <input message="tns:sayHello"/>
+            <output message="tns:sayHelloResponse"/>
+        </operation>
+    </portType>
+    <binding name="UsernameTokenPortBinding" type="tns:ServiceIface">
+<!--        <wsp:PolicyReference URI="#UsernameTokenPolicy"/>-->
+        <soap:binding transport="http://schemas.xmlsoap.org/soap/http" style="document"/>
+        <operation name="sayHello">
+            <soap:operation soapAction=""/>
+            <input>
+                <soap:body use="literal"/>
+            </input>
+            <output>
+                <soap:body use="literal"/>
+            </output>
+        </operation>
+    </binding>
+    <service name="UsernameToken">
+        <port name="UsernameTokenPort" binding="tns:UsernameTokenPortBinding">
+            <soap:address location="http://@jboss.bind.address@:8080/UsernameTokenElytronTestCase/UsernameToken/ElytronUsernameTokenImpl"/>
+        </port>
+    </service>
+
+<!--    <wsp:Policy wsu:Id="UsernameTokenPolicy">-->
+<!--        <wsp:ExactlyOne>-->
+<!--            <wsp:All>-->
+<!--                <sp:SupportingTokens xmlns:sp="http://schemas.xmlsoap.org/ws/2005/07/securitypolicy">-->
+<!--                    <wsp:Policy>-->
+<!--                        <sp:UsernameToken sp:IncludeToken="http://schemas.xmlsoap.org/ws/2005/07/securitypolicy/IncludeToken/AlwaysToRecipient">-->
+<!--                            <wsp:Policy>-->
+<!--                                <sp:WssUsernameToken10/>-->
+<!--                            </wsp:Policy>-->
+<!--                        </sp:UsernameToken>-->
+<!--                    </wsp:Policy>-->
+<!--                </sp:SupportingTokens>-->
+<!--            </wsp:All>-->
+<!--        </wsp:ExactlyOne>-->
+<!--    </wsp:Policy>-->
+
+</definitions>
