--- integration-tests/cxf-soap/src/test/resources/__files/soapObjectResponse.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0"?>
-<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/"
-                  xmlns:hello="http://www.helloworld.com/Service/">
-    <soapenv:Header/>
-    <soapenv:Body>
-        <ns2:PersonResponse xmlns:ns2="http://www.helloworld.com/Service/">
-            <firstName>Camel Quarkus</firstName>
-            <lastName>CXF</lastName>
-            <greeting>Hello</greeting>
-        </ns2:PersonResponse>
-    </soapenv:Body>
-</soapenv:Envelope>
